# projects/gin-advanced/docs/swagger.yaml
basePath: /
definitions:
...
  api.LoginRequest:
    properties:
      captcha:
        type: string
      captcha_id:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - captcha
    - captcha_id
    - password
    - username
    type: object
  api.LoginResponse:
    properties:
      nickname:
        type: string
      permissions:
        items:
          type: string
        type: array
      roles:
        items:
          type: string
        type: array
      token:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  api.RegisterRequest:
    properties:
      captcha:
        type: string
      captcha_id:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - captcha
    - captcha_id
    - password
    - username
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Permission:
    properties:
      created_at:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 权限描述
        type: string
      id:
        type: integer
      method:
        description: 请求方法 GET/POST/PUT/DELETE
        type: string
      name:
        description: 权限名称，如 "user:create"
        type: string
      path:
        description: 请求路径
        type: string
      updated_at:
        type: integer
    type: object
  models.Role:
    properties:
      created_at:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 角色描述
        type: string
      id:
        type: integer
      name:
        description: 角色名称
        type: string
      permissions:
        description: 权限关联
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: integer
    type: object
  models.User:
    properties:
      created_at:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      nickname:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: integer
      username:
        type: string
    type: object
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Gin-Advanced RESTful API 文档
  title: Gin-Advanced API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: 用户登录接口，需要用户名、密码和验证码，返回JWT token
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 用户登录
      tags:
      - 用户
  /register:
    post:
      consumes:
      - application/json
      description: 用户注册接口，需要提供用户名、密码和验证码
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 用户注册
      tags:
      - 用户
...
  /user/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 用户
  /user/info/{id}:
    get:
      consumes:
      - application/json
      description: 根据用户ID获取用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取用户信息
      tags:
      - 用户
  /user/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定用户的详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - 用户
securityDefinitions:
  BearerAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"